21_mock_interviewer_multi_agent_using_crewai/ai_mock_interviewer/pyproject.toml
[tool.poetry]
name = "ai_mock_interviewer"
version = "0.1.0"
description = "AI Mock Interviewer based on CrewAI"
authors = ["Name <email>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
crewai = {extras = ["tools"], version = "^0.114.0"}
streamlit = "^1.44.1"
streamlit-mic-recorder = "^0.0.8"
speechrecognition = "^3.14.2"
openai-whisper = "^20240930"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.9.3"
pdbp = "^1.6.1"

[tool.ruff]
extend = "ruff.toml"